# Install collective.indexing