--- a/docs/HISTORY.txt
+++ b/docs/HISTORY.txt
@@ -5,6 +5,9 @@ Changelog
 4.14.2 (unreleased)
 -------------------
 
+- Install collective.indexing.
+  [mathias.leimgruber]
+
 - Protect date of cassation and submission with a separate edit permission.
   Only managers should be allowed to edit those dates.
   [phgross]