local performance measurements. 

**Test-Case: Creating 40 different files incl. tika.**
WITHOUT: 37s
WITH: 21s


**Test-Case: 5 parallel users uploading files with quickupload**
WITHOUT: 2s pro Request
WITH: 3.3s pro Request

**Test-Case: 1 seriell users uploading files with quickupload**
WITHOUT: 1 rps (880ms per Request)
WITH: 1.6 rps (880ms per Request)