@maethu 2017-01-06
c.indexing ist im Moment bei uns auf 62 (Aus der Matrix) Deployments installiert. 
Darunter iZug / zg.ch / bern.ch IntranetBern / stadt.winterthur.ch und weitere. 
So weit ich das beurteilen kann, läuft das ganze sehr stabil. 


`c.indexing` patched im wesentlichen `index`, `unindex` und `reindex` vom portal catalog. 
Anstatt direkt den catalog zu verwenden wird bei jeder operation ein Job gequeued. 

In einem before commit hook werden danach folgende Operationen ausgeführt. 

1. die Queue wird optimiert. z.B Falls ein Obj in einem Request 3 mal `reindex` wird, wird dies auf einmal reduziert. 
2. Die Operationen werden auf jedem registrierten `ICatalog...` utility  ausgeführt. Nach dem installieren vom c.indexing wird ein standtad `Icatalog...` registriert. collective.solr z.B registriert einen zweien ICatalog. 
3. Jedes mal wenn ein ICatalog abgefragt wird, wird die Queue erst geleert, bzw. abgearbeitet. 


Mögliche Probleme:
1. Catalog queries während dem Abarbeiten der Queue können zu einem inkonsistenten State führen. 
2. Hatte früher zu Problemen geführt, z.b mit ftw.copymovepatches. 
3. Verlangsamt unter umständen mit collective.solr zusammen. das indizieren. Es werden immer 2 ICatallog indiziert. In der Regel wird die Performance, aber besser da die Queue optimiert wird.