@lukasgraf 2017-01-06
@maethu wie am FedEx-Day besprochen: Aus meiner Sicht grundsätzlich 👍 ! Da es keine Test-Failures gibt erwarte ich eigentlich keine grossen Probleme, und sonst müssen wir sie halt angehen.

2x indizieren in Kombination mit `collective.solr` verstehe ich jetzt nicht ganz: `c.indexing` ist ja sowieso eine harte dependency von `c.solr`, oder nicht mehr?

@deiferni @phgross aus meiner Sicht spricht nichts dagegen, das mal zu mergen. Wie seht ihr das?