@maethu 2017-01-06
> 2x indizieren in Kombination mit collective.solr verstehe ich jetzt nicht ganz: c.indexing ist ja sowieso eine harte dependency von c.solr, oder nicht mehr?

@lukasgraf Ist nur ein hint, falls später noch `c.solr` dazu kommt. Im Moment wird nur einmal indiziert, da nur ein ICatalog Utility vorhanden ist. `c.solr`, bzw. `ftw.solr` wird mit diesem PR nicht installiert. Es geht ja um die "reine" Performance-Verbesserung mit `c.indexing`